Minor code cleanup in Hast.Samples.Consumer.
